perf: switch partial reads to use bitwise AND
This switches partial `read` syscalls (i.e. those that don't read
multiples of 8 bytes) to use bitwise AND operators instead of MUL and
DIV when combining values bytewise. The goal is to produce a simpler
formula for the solver. Note that incomplete solvers in particular have
a hard time solving "t = x * (2^N) / (2^N)" for example.